Fix projection column types: int8→Int64, JSON→JSON::Any
The ES::ProjectionDSL macro maps Crystal types to PostgreSQL types:
- `int8` is not a recognized Crystal type and fell through to TEXT
- `JSON` similarly maps to TEXT; only `JSON::Any` produces JSONB

Replace all `int8` with `Int64` (→ BIGINT) and all `JSON` with `JSON::Any`
(→ JSONB) across every projection's define_projection block.

+++ app/src/domains/api_keys/projections/api_keys.cr
@@ -6,7 +6,7 @@ module CrystalBank::Domains::ApiKeys
       define_projection "projections.api_keys", init: true do
         column :id, Int32, serial: true, primary_key: true
         column :uuid, UUID, null: false
-        column :aggregate_version, int8, null: false
+        column :aggregate_version, Int64, null: false
         column :scope_id, UUID, null: false
         column :created_at, Time, null: false
         column :name, String, null: false
